Fix %llu truncation: use unsigned long long for uint64_t DOE CSV columns
On aarch64, mixing uint32_t and uint64_t arguments through va_arg with %lu
causes the 64-bit values to be printed as 32-bit (elapsed_ns truncated:
14,038,220,000 → 1,153,318,112 = value mod 2^32). riscv64 and amd64 were
unaffected because either values fit in 32 bits or the va_arg ABI aligns
differently.

Fix: add %llu handler to kvsnprintf (shim.c) using unsigned long long, and
change doe_log.c to cast all uint64_t arguments to (unsigned long long) and
use %llu. unsigned long long is unambiguously 64-bit on all platforms.

+++ src/starkernel/doe_log.c
@@ -83,25 +83,27 @@ void doe_log_tick_row(VM *vm, const HeartbeatTickSnapshot *snap)
     jitter_bits.d = snap->estimated_jitter_ns;
     uint64_t jitter_raw = jitter_bits.u;
 
-    /* Format the 15-column CSV row into a local buffer */
+    /* Format the 15-column CSV row into a local buffer.
+     * Use %llu (unsigned long long) for all uint64_t fields — %lu is unreliable
+     * for values > 2^32 on aarch64 due to mixed-width varargs ABI behaviour. */
     char buf[DOE_BUF_SIZE];
     snprintf(buf, sizeof(buf),
-             "%u,%lu,%lu,%u,%u,%u,%lu,%lu,%u,%u,%u,%lu,%lu,%lu,%lu",
+             "%u,%llu,%llu,%u,%u,%u,%llu,%llu,%u,%u,%u,%llu,%llu,%llu,%llu",
              snap->tick_number,
-             snap->elapsed_ns,
-             snap->tick_interval_ns,
+             (unsigned long long)snap->elapsed_ns,
+             (unsigned long long)snap->tick_interval_ns,
              snap->cache_hits_delta,
              snap->bucket_hits_delta,
              snap->word_executions_delta,
-             snap->hot_word_count,
-             avg_heat_q48,
+             (unsigned long long)snap->hot_word_count,
+             (unsigned long long)avg_heat_q48,
              snap->window_width,
              snap->actual_window_size,
              snap->predicted_label_hits,
-             jitter_raw,
-             apic_ticks,
-             time_trust_q48,
-             variance_q48);
+             (unsigned long long)jitter_raw,
+             (unsigned long long)apic_ticks,
+             (unsigned long long)time_trust_q48,
+             (unsigned long long)variance_q48);
 
     console_puts(DOE_PREFIX);
     console_puts(buf);
